refactor(executor): extract redirection logic into separate module
- Created src/executor/redirection.rs with ~600 lines of redirection code
- Moved all redirection functions from executor/mod.rs to new module
- Added comprehensive module-level documentation explaining:
  - Purpose of the redirection module
  - Different redirection types supported
  - File descriptor table integration
  - Noclobber support
- Updated executor/mod.rs with proper module declarations and imports
- All 554 tests pass successfully
- Part of Phase 1 executor refactoring (Step 1.2)
